Fix issue caused by passing frozen string to pyu-ruby-sasl
Omniauth freezes configuration parameters that come in from the initializer.
Then, omniauth-ldap extracts those params and passes them to pyu-ruby-sasl as a
part of the challenge response stage.

Unfortuantely pyu-ruby-sasl does some in place string manipulation that fails
with an exception if performed on a frozen string. Theres no straightforward way
to unfreeze a string in ruby, so just .dup it before sending it down to pyu-ruby-sasl.