Escape username before building ldap filter. (#1)
Certain characters are valid for usernames, but must be escaped before a filter
is built with the username. Otherwise, these chars get interpreted as special
filter chars which can do anything from break the filter creation to enable an
injection attack to build a malicious filter. These special chars include '*',
'(', ')', and '\'.

ruby-net-ldap already has escaping implemented, but the escape method isn't
currently called because omniauth-ldap passes a full filter string in. The
escape method needs to be called before omniauth-ldap interpolates the user
supplied username into the configured filter template.
This is only relevant when there is a filter template specified.

See RFC 2254, "String Representation of LDAP" for more information.